requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cloudinary==1.40.0
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
